Refactor setting the global JRuby ScriptingContainer up.
